+++ models/MongoConfig.ts
@@ -1,5 +1,5 @@
 import { Config } from "./Config";
-import { ConfigDB, MongoConfig } from "../mongo"
+import { ConfigDB, MongoConfig } from "../mongo/models/MongoConfig"
 
 export interface MongoDBConfig extends Config {
 	db: ConfigDB,