+++ HISTORY.md
@@ -1,4 +1,12 @@
 ## [Unreleased]
+### Fixed
+- cfg property of MongoApp was not correct
+### Changed
+- Config.master renamed to app
+### Added
+- Config.log.devel
+### Removed
+- Config.worker
 
 ## [0.1.0] - 2019-02-22
 ### Initial commit